one route two methods

+++ api/v1/views/places.py
@@ -24,22 +24,21 @@ def get_places(city_id):
                  methods=["GET"])
 def a_place(place_id):
     """retrieves a place"""
-    place = storage.get(Place, place_id)
-    if not place:
-        abort(404)
-    return jsonify(place.to_dict())
-
-
-@app_views.route("/places/<place_id>", strict_slashes=False,
-                 methods=["DELETE"])
-def delete_place(place_id):
-    """deletes a place"""
-    place = storage.get(Place, place_id)
-    if not place:
-        abort(404)
-    storage.delete(place)
-    storage.save()
-    return jsonify({}), 200
+    if request.method == "GET":
+        place = storage.get(Place, place_id)
+        if not place:
+            abort(404)
+        return jsonify(place.to_dict())
+# @app_views.route("/places/<place_id>", strict_slashes=False,
+#                 methods=["DELETE"])
+# def delete_place(place_id):
+    if request.method == "DELETE":
+        place = storage.get(Place, place_id)
+        if not place:
+            abort(404)
+        storage.delete(place)
+        storage.save()
+        return jsonify({}), 200
 
 
 @app_views.route("/cities/<city_id>/places", strict_slashes=False,
